requirements.txt
Flask
google-generativeai
Pillow
python-dotenv
